Major refactor - pruning of warnings
- I have been working on the project for a few months at this point,
  on release mode. So I have not seen a lot of warnings in the project
  for a while. I enabled debug mode and focused on all warnings

- Even though the main intention was to solve meaningful warnings,
  I ended up solving many bugs in the code

- One of the key additions in the updates is `std::unreachable`. I
  migrated the project to C++23 just for this feature

- Some of the common warnings were related to unused arguments and
  unused values

- Many includes have been reorganised. Previously there were includes
  in one file that were not used by the file itself, but instead by
  other files in the sub-hierarchy. This has been fixed and includes
  are now more relevant to the file they are present in

- Some minor and major bugfixes

+++ src/utils/utils.hpp
@@ -38,7 +38,7 @@ useit inline String number_to_position(u64 number) {
 
 useit inline Vec<String> split_string(const String& value, const String& slice) {
 	Vec<String> result;
-	int         index = 0;
+	usize       index = 0;
 	while (index < value.length()) {
 		auto pos = value.find(slice, index);
 		if (pos != String::npos) {
